feat: Se ajusta seguimiento a pagos mensuales creando pago pago_mensual_auditoria y modificando la transaccion masiva de pagos
