+++ models/pago_mensual.go
@@ -7,15 +7,8 @@ import (
 	"strings"
 	"time"
 
-	"github.com/astaxie/beego/logs"
 	"github.com/astaxie/beego/orm"
 )
-
-type PagoMensualAuditoria struct {
-	Pago              *PagoMensual
-	CargoEjecutor     string
-	DocumentoEjecutor string
-}
 
 type PagoMensual struct {
 	Id                     int                `orm:"column(id);pk;auto"`
@@ -166,76 +159,6 @@ func DeletePagoMensual(id int) (err error) {
 		var num int64
 		if num, err = o.Delete(&PagoMensual{Id: id}); err == nil {
 			fmt.Println("Number of records deleted in database:", num)
-		}
-	}
-	return
-}
-
-// funcion para agregar un cambio de estado pago de manera invisible para llevar una auditoria
-func seguimientoAuditoria(m *PagoMensualAuditoria) (v CambioEstadoPago) {
-	v = CambioEstadoPago{
-		EstadoPagoMensualId:    m.Pago.EstadoPagoMensualId.Id,
-		DocumentoResponsableId: m.DocumentoEjecutor,
-		CargoResponsable:       m.CargoEjecutor,
-		PagoMensualId:          m.Pago,
-		Activo:                 true,
-		FechaCreacion:          time.Time{},
-		FechaModificacion:      time.Time{},
-	}
-	return v
-}
-
-func UpdatePagoMensualAuditoriaById(m *PagoMensualAuditoria) (err error) {
-	o := orm.NewOrm()
-	err = o.Begin()
-
-	defer func() {
-		if r := recover(); r != nil {
-			logs.Error(r)
-			o.Rollback()
-		} else {
-			o.Commit()
-		}
-
-	}()
-
-	v := PagoMensual{Id: m.Pago.Id}
-	// ascertain id exists in the database
-	if err = o.Read(&v); err == nil {
-		if _, err = o.Update(m.Pago); err != nil {
-			panic(err)
-		} else {
-			s := seguimientoAuditoria(m)
-			if _, err := o.Insert(&s); err != nil {
-				panic(err)
-			}
-		}
-	}
-	return
-}
-
-func AddPagoMensualAuditoria(m *PagoMensualAuditoria) (id int64, err error) {
-	o := orm.NewOrm()
-	err = o.Begin()
-	defer func() {
-		if r := recover(); r != nil {
-			logs.Error(r)
-			o.Rollback()
-		} else {
-			o.Commit()
-		}
-
-	}()
-
-	if id, err = o.Insert(m.Pago); err != nil {
-		logs.Error(err)
-		panic(err)
-	} else {
-		v := seguimientoAuditoria(m)
-
-		if id, err = o.Insert(&v); err != nil {
-			logs.Error(err)
-			panic(err)
 		}
 	}
 	return
