+++ models/tr_aprobacion_masiva_pagos.go
@@ -1,30 +1,64 @@
 package models
 
 import (
-	"fmt"
+	"time"
 
+	"github.com/astaxie/beego/logs"
 	"github.com/astaxie/beego/orm"
 )
 
-//funcion para la aprobación masiva de pagos
-func AprobarPagos(m *[]PagoMensual) (err error) {
-	o := orm.NewOrm()
+type ArregloPagoMensualAuditoria struct {
+	Pagos             *[]PagoMensual
+	CargoEjecutor     string
+	DocumentoEjecutor string
+}
 
+//funcion para la aprobación masiva de pagos
+func AprobarPagos(m *ArregloPagoMensualAuditoria) (err error) {
+	o := orm.NewOrm()
 	err = o.Begin()
 	if err != nil {
-		fmt.Println(err)
+		logs.Error(err)
+		panic(err)
 	}
 
-	for _, v := range *m {
+	defer func() {
+		if r := recover(); r != nil {
+			logs.Error(r)
+			o.Rollback()
+			outputError := map[string]interface{}{"funcion": "/AprobarPagos", "err": err, "status": "500"}
+			panic(outputError)
+
+		} else {
+			err = o.Commit()
+			return
+		}
+
+	}()
+
+	for _, v := range *m.Pagos {
 		v.EstadoPagoMensualId.Id = 5
 		if _, err = o.Update(&v); err != nil {
-			fmt.Println("Pago mensual, pago aprobado: ", &v)
-			err = o.Rollback()
+			panic(err)
 		} else {
-			fmt.Println(err)
+			s := seguimientoAuditoria(&v, m.CargoEjecutor, m.DocumentoEjecutor)
+			/*if _, err := o.Insert(&s); err != nil {
+				panic(err)
+			}*/
+			// Se cambian los estados anteriores activos
+			if _, err := o.QueryTable("cambio_estado_pago").Filter("pago_mensual_id", v.Id).Filter("activo", true).Update(orm.Params{
+				"activo":             false,
+				"fecha_modificacion": time.Now(),
+			}); err != nil {
+				panic(err)
+			} else {
+				if _, err := o.Insert(&s); err != nil {
+					panic(err)
+				}
+			}
+
 		}
 	}
-	err = o.Commit()
 
 	return
 }
